transforms/code/license_copyright_removal/ray/requirements.txt
# transform runtime
#data-prep-kit==0.1.2
# Transform requirements except data-prep-kit (see above)
scancode-toolkit==32.1.0
